+++ derby.log
@@ -1,0 +1,13 @@
+----------------------------------------------------------------
+Wed Jul 31 10:16:37 EDT 2019:
+Booting Derby version The Apache Software Foundation - Apache Derby - 10.12.1.1 - (1704137): instance a816c00e-016c-4863-d2a4-00000ba4cb08 
+on database directory memory:/Users/alexp/Desktop/My projects/MS3/MS3_BootCamp/testdb with class loader sun.misc.Launcher$AppClassLoader@18b4aac2 
+Loaded from file:/Users/alexp/.m2/repository/org/apache/derby/derby/10.12.1.1/derby-10.12.1.1.jar
+java.vendor=Oracle Corporation
+java.runtime.version=1.8.0_221-b11
+user.dir=/Users/alexp/Desktop/My projects/MS3/MS3_BootCamp
+os.name=Mac OS X
+os.arch=x86_64
+os.version=10.14.6
+derby.system.home=null
+Database Class Loader started - derby.database.classpath=''
