+++ src/main/resources/application.properties
@@ -1,1 +1,6 @@
 server.port=8080
+
+#spring.datasource.url=jdbc:mysql://localhost/starTrackShips?useLegacyDatetimeCode=false&serverTimezone=UTC
+#spring.datasource.username=root
+#spring.datasource.password=rootroot
+#spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
